replace double with BigDecimal

+++ backend/src/main/java/enteties/model/Point.java
@@ -2,14 +2,16 @@ package enteties.model;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 /**
  * Класс-абстракция над точкой на координатной плоскости.
  * Нужен для обработки данных на бэкенде.
  */
 @Getter
 public class Point {
-    private final double x;
-    private final double y;
+    private final BigDecimal x;
+    private final BigDecimal y;
 
     /**
      * Конструктор класса Point
@@ -17,7 +19,7 @@ public class Point {
      * @param x значение по оси x.
      * @param y значение по оси y.
      */
-    public Point(double x, double y) {
+    public Point(BigDecimal x, BigDecimal y) {
         this.x = x;
         this.y = y;
     }
